Add `background.js` to fetch the current tab origin
With the previous implementation, where the content.js was responsible for returning the origin, it was quite common that we get the following error: Could not establish connection. Receiving end does not exist.
This because the content.js wasn't ready for receiving messages.

With the background.js approach, we hopefully have a reliable way of determine the origin of the current active tab.